+++ sdks/python/apache_beam/runners/__init__.py
@@ -20,10 +20,10 @@
 This package defines runners, which are used to execute a pipeline.
 """
 
-from apache_beam.runners.dataflow_runner import DataflowRunner
 from apache_beam.runners.direct.direct_runner import DirectRunner
 from apache_beam.runners.direct.direct_runner import EagerRunner
-from apache_beam.runners.runner import create_runner
+from apache_beam.runners.google_cloud_dataflow.dataflow_runner import DataflowRunner
 from apache_beam.runners.runner import PipelineRunner
 from apache_beam.runners.runner import PipelineState
+from apache_beam.runners.runner import create_runner
 from apache_beam.runners.test.test_dataflow_runner import TestDataflowRunner
