+++ sdks/python/apache_beam/runners/runner.py
@@ -35,7 +35,8 @@ def _get_runner_map(runner_names, module_path):
 
 
 _DIRECT_RUNNER_PATH = 'apache_beam.runners.direct.direct_runner.'
-_DATAFLOW_RUNNER_PATH = 'apache_beam.runners.dataflow_runner.'
+_DATAFLOW_RUNNER_PATH = (
+    'apache_beam.runners.google_cloud_dataflow.dataflow_runner.')
 _TEST_RUNNER_PATH = 'apache_beam.runners.test.'
 
 _KNOWN_DIRECT_RUNNERS = ('DirectRunner', 'EagerRunner')
