+++ sdks/python/apache_beam/runners/template_runner_test.py
@@ -20,14 +20,14 @@
 from __future__ import absolute_import
 
 import json
-import unittest
 import tempfile
+import unittest
 
 import apache_beam as beam
-from apache_beam.pipeline import Pipeline
-from apache_beam.runners.dataflow_runner import DataflowRunner
-from apache_beam.utils.pipeline_options import PipelineOptions
 from apache_beam.internal import apiclient
+from apache_beam.pipeline import Pipeline
+from apache_beam.runners.google_cloud_dataflow.dataflow_runner import DataflowRunner
+from apache_beam.utils.pipeline_options import PipelineOptions
 
 
 class TemplatingDataflowRunnerTest(unittest.TestCase):
